Release/v0.3.0

+++ packages/primitive-contracts/contracts/option/interfaces/IFlash.sol
@@ -1,19 +1,11 @@
 // SPDX-License-Identifier: MIT
 
-
-
-
-
-
-
-
-
 pragma solidity ^0.6.2;
 
 interface IFlash {
     function primitiveFlash(
         address receiver,
-        uint256 outUnderlyings,
+        uint outUnderlyings,
         bytes calldata data
     ) external;
 }
