+++ packages/primitive-contracts/contracts/option/interfaces/IRedeem.sol
@@ -1,13 +1,5 @@
 // SPDX-License-Identifier: MIT
 
-
-
-
-
-
-
-
-
 pragma solidity ^0.6.2;
 
 interface IRedeem {
@@ -17,7 +9,7 @@ interface IRedeem {
 
     function factory() external view returns (address);
 
-    function mint(address user, uint256 amount) external;
+    function mint(address user, uint amount) external;
 
-    function burn(address user, uint256 amount) external;
+    function burn(address user, uint amount) external;
 }
